pedidos e historico em andamento

+++ src/main/java/com/br/api/lanchonete/domain/product/ProductUpdateDTO.java
@@ -2,11 +2,12 @@ package com.br.api.lanchonete.domain.product;
 
 import org.springframework.web.multipart.MultipartFile;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 public record ProductUpdateDTO(
         Optional<String> name,
-        Optional<Double> price,
+        Optional<BigDecimal> price,
         Optional<Category> category,
         Optional<String> description,
         Optional<Integer> stock,
